+++ ANDROID/TravlToTibet/app/src/main/java/com/dean/travltotibet/adapter/TeamRequestListAdapter.java
@@ -114,7 +114,7 @@ public class TeamRequestListAdapter extends BaseAdapter {
             holder.mUserIcon.setImageResource(R.drawable.gray_profile);
         }
 
-        holder.mWatch.setText(request.getWatch() + "");
+        holder.mWatch.setText(request.getWatch().intValue() + "");
 
         holder.rippleLayout.setOnClickListener(new View.OnClickListener() {
             @Override
